chore: reorganise store
